backend/requirements.txt
Flask
Flask-Cors
flask-sock
google-cloud-speech
google-generativeai